move duplicate code to utilities
