+++ src/js/svgo-worker/index.js
@@ -1,4 +1,5 @@
 import { optimize } from 'svgo/dist/svgo.browser.js';
+import { collectPlugins } from '../utils/settings.js';
 
 const createDimensionsExtractor = () => {
   const dimensions = {};
@@ -34,30 +35,7 @@ const createDimensionsExtractor = () => {
 
 function compress(svgInput, settings) {
   // setup plugin list
-  const floatPrecision = Number(settings.floatPrecision);
-  const transformPrecision = Number(settings.transformPrecision);
-  const plugins = [];
-
-  for (const [name, enabled] of Object.entries(settings.plugins)) {
-    if (!enabled) continue;
-
-    const plugin = {
-      name,
-      params: {},
-    };
-
-    // TODO: revisit this
-    // 0 almost always breaks images when used on `cleanupNumericValues`.
-    // Better to allow 0 for everything else, but switch to 1 for this plugin.
-    plugin.params.floatPrecision =
-      plugin.name === 'cleanupNumericValues' && floatPrecision === 0
-        ? 1
-        : floatPrecision;
-
-    plugin.params.transformPrecision = transformPrecision;
-
-    plugins.push(plugin);
-  }
+  const plugins = collectPlugins(settings);
 
   // multipass optimization
   const [dimensions, extractDimensionsPlugin] = createDimensionsExtractor();
