+++ src/js/utils/settings.js
@@ -1,0 +1,28 @@
+export function collectPlugins(settings) {
+  const floatPrecision = Number(settings.floatPrecision);
+  const transformPrecision = Number(settings.transformPrecision);
+  const plugins = [];
+
+  for (const [name, enabled] of Object.entries(settings.plugins)) {
+    if (!enabled) continue;
+
+    const plugin = {
+      name,
+      params: {},
+    };
+
+    // TODO: revisit this
+    // 0 almost always breaks images when used on `cleanupNumericValues`.
+    // Better to allow 0 for everything else, but switch to 1 for this plugin.
+    plugin.params.floatPrecision =
+      plugin.name === 'cleanupNumericValues' && floatPrecision === 0
+        ? 1
+        : floatPrecision;
+
+    plugin.params.transformPrecision = transformPrecision;
+
+    plugins.push(plugin);
+  }
+
+  return plugins;
+}
